create usersController
